feat: add `Client#subscribe`
The purpose of this addition is to be able to provide Pub/Sub support for `PoolClient` and `ClusterClient` in the future.
