+++ command.ts
@@ -10,7 +10,7 @@ import type {
   SimpleString,
 } from "./protocol/shared/types.ts";
 import type { RedisPipeline } from "./pipeline.ts";
-import type { RedisSubscription } from "./pubsub.ts";
+import type { RedisSubscription } from "./client.ts";
 import type {
   StartEndCount,
   XAddFieldValues,
