+++ pool/client.ts
@@ -1,7 +1,14 @@
 import type { Connection, SendCommandOptions } from "../connection.ts";
 import type { Pool } from "./pool.ts";
-import type { Client } from "../client.ts";
+import type {
+  Client,
+  DefaultPubSubMessageType,
+  PubSubMessageType,
+  RedisSubscription,
+  SubscribeCommand,
+} from "../client.ts";
 import { DefaultClient } from "../client.ts";
+import { NotImplementedError } from "../errors.ts";
 import type { RedisReply, RedisValue } from "../protocol/shared/types.ts";
 
 export function createPoolClient(pool: Pool<Connection>): Client {
@@ -43,6 +50,13 @@ class PoolClient implements Client {
     } finally {
       this.#pool.release(connection);
     }
+  }
+
+  subscribe<TMessage extends PubSubMessageType = DefaultPubSubMessageType>(
+    command: SubscribeCommand,
+    ...channelsOrPatterns: Array<string>
+  ): Promise<RedisSubscription<TMessage>> {
+    return Promise.reject(new NotImplementedError("PoolClient#subscribe"));
   }
 
   close(): void {
